[Code Refactoring] Separate Out Create New Modal Component Implementation to a New File #9
[Code Refactoring] Separate Out Create New Modal Component Implementation to a New File #9
