+++ frontend/src/components/ownerDetails/OwnerDetailsLoggedInView.tsx
@@ -5,6 +5,7 @@ import * as UserModel from "../../models/user";
 import * as OwnerDetailsModel from "../../models/ownerDetails";
 import ownerDetailsPageStyle from "../../styles/OwnerDetailsPage.module.css";
 import * as commonImports from "../../commonCode/importMRTRelated";
+import {CreateNewOwnerDetailsModal} from "./commonElement/CreateNewOwnerDetailsModal";
 
 //Strategy DesignPattern Used for the Create, Update and Delete Operations
 import { CreateNewRowStrategy } from "./commonElement/Strategy/CreateNewRowStrategy";
@@ -166,177 +167,20 @@ const OwnerDetailsLoggedInView = () => {
             </commonImports.Button>
           )}
         />
-        <CreateNewAccountModal
+        <CreateNewOwnerDetailsModal
           columns={ownerDetailsGridColumns}
           open={createModalOpen}
           onClose={() => setCreateModalOpen(false)}
           onSubmit={handleCreateNewRow}
+          usersArr={usersArr}
         />
       </commonImports.Container>
     </>
   );
 };
 
-interface CreateModalProps {
-  columns: commonImports.MRT_ColumnDef<OwnerDetailsModel.IOwnerDetailsViewModel>[];
-  onClose: () => void;
-  onSubmit: (values: OwnerDetailsModel.IOwnerDetailsViewModel) => void;
-  open: boolean;
-}
 
-//example of creating a mui dialog modal for creating new rows
-export const CreateNewAccountModal = ({
-  open,
-  columns,
-  onClose,
-  onSubmit,
-}: CreateModalProps) => {
-  const [values, setValues] = commonImports.useState<any>(() =>
-    columns.reduce((acc, column) => {
-      acc[column.accessorKey ?? ""] = "";
-      return acc;
-    }, {} as any)
-  );
 
-  const [selectedOwnerType, setSelectedOwnerType] = commonImports.useState("");
-  const [errors, setErrors] = commonImports.useState<{ [key: string]: string }>(
-    {}
-  );
-
-  const validate = () => {
-    let tempErrors = {};
-    tempErrors = {
-      ...tempErrors,
-      userId: values.userId ? "" : "This field is required",
-      ownerEmail: values.ownerEmail
-        ? /\S+@\S+\.\S+/.test(values.ownerEmail)
-          ? ""
-          : "Email is not valid"
-        : "This field is required",
-      ownerWebsite: values.ownerWebsite
-        ? /^(ftp|http|https):\/\/[^ "]+$/.test(values.ownerWebsite)
-          ? ""
-          : "URL is not valid"
-        : "This field is required",
-      ownerName: values.ownerName ? "" : "This field is required",
-      ownerMobileNo: values.ownerMobileNo
-        ? /^[1-9][0-9]{9}$/.test(values.ownerMobileNo)
-          ? ""
-          : "Please enter a valid mobile number"
-        : "This field is required",
-    };
-    setErrors({
-      ...tempErrors,
-    });
-
-    return Object.values(tempErrors).every((x) => x === "");
-  };
-
-  const handleSubmit = () => {
-    //put your validation logic here
-    if (validate()) {
-      onSubmit(values);
-      onClose();
-    }
-  };
-
-  return (
-    <commonImports.Dialog open={open}>
-      <commonImports.DialogTitle textAlign="center">
-        Create New Owner
-      </commonImports.DialogTitle>
-      <commonImports.DialogContent>
-        <form onSubmit={(e) => e.preventDefault()}>
-          <commonImports.Stack
-            sx={{
-              width: "100%",
-              minWidth: { xs: "300px", sm: "360px", md: "400px" },
-              gap: "1.5rem",
-            }}
-          >
-            {columns
-              .filter((column) => column.accessorKey === "userId")
-              .map((column) => (
-                <commonImports.FormControl
-                  error={column.accessorKey && !!errors[column.accessorKey]}
-                >
-                  <commonImports.Select
-                    label="User Id"
-                    key={column.accessorKey}
-                    name={column.accessorKey}
-                    value={selectedOwnerType}
-                    onChange={(event) => {
-                      setValues({
-                        ...values,
-                        [event.target.name]: event.target.value,
-                      });
-                      setSelectedOwnerType(event.target.value);
-                    }}
-                    displayEmpty
-                    sx={{ minWidth: 120 }}
-                  >
-                    <commonImports.MenuItem value="" disabled>
-                      Select a User
-                    </commonImports.MenuItem>
-                    {usersArr.map((option) => (
-                      <commonImports.MenuItem
-                        key={option._id}
-                        value={option._id}
-                      >
-                        {option.username + " - " + option.role}
-                      </commonImports.MenuItem>
-                    ))}
-                  </commonImports.Select>
-                  <commonImports.FormHelperText>
-                    {column.accessorKey &&
-                      errors.hasOwnProperty(column.accessorKey)
-                      ? errors[column.accessorKey]
-                      : ""}
-                  </commonImports.FormHelperText>
-                </commonImports.FormControl>
-              ))}
-
-            {columns
-              .filter(
-                (column) =>
-                  column.accessorKey !== "_id" &&
-                  column.accessorKey !== "userId" &&
-                  column.accessorKey !== "createdAt" &&
-                  column.accessorKey !== "updatedAt"
-              )
-              .map((column) => (
-                <commonImports.TextField
-                  key={column.accessorKey}
-                  label={column.header}
-                  name={column.accessorKey}
-                  onChange={(e) =>
-                    setValues({ ...values, [e.target.name]: e.target.value })
-                  }
-                  error={column.accessorKey && !!errors[column.accessorKey]}
-                  helperText={
-                    column.accessorKey &&
-                      errors.hasOwnProperty(column.accessorKey)
-                      ? errors[column.accessorKey]
-                      : ""
-                  }
-                />
-              ))}
-          </commonImports.Stack>
-        </form>
-      </commonImports.DialogContent>
-      <commonImports.DialogActions sx={{ p: "1.25rem" }}>
-        <commonImports.Button onClick={onClose}>Cancel</commonImports.Button>
-        <commonImports.Button
-          color="secondary"
-          onClick={handleSubmit}
-          variant="contained"
-        >
-          Create New Owner
-        </commonImports.Button>
-      </commonImports.DialogActions>
-    </commonImports.Dialog>
-  );
-};
 
 export default OwnerDetailsLoggedInView;
